Crate gpiocdev_uapi

A thin but safe Rust layer around the Linux GPIO uAPI.

Modules

v1

This module implements GPIO ABI v1 which was released in Linux v4.8.

v2

This module implements GPIO ABI v2 which is the current version of the ABI, released in Linux v5.10.

Structs

Name

A uAPI name string, common to ABI v1 and v2.

ValidationError

A failure to validate a struct returned from a system call.

Enums

Error

Errors returned by gpiocdev_uapi functions.

Constants

NAME_LEN_MAX

The maximum number of bytes stored in a Name.

NUM_LINES_MAX

The maximum number of lines that may be requested in a single request.

Functions

has_event

Check if the file has an event available to read.

read_event

Read an event from a chip or request file descriptor.

wait_event

Wait for the file to have an event available to read.

Type Definitions

Result

The result returned by gpiocdev_uapi functions.