Module gpiocdev_uapi::v1

This module implements GPIO ABI v1 which was released in Linux v4.8.

This ABI version is deprecated.

ABI v2 adds features and overcomes a number of bugs and limitations present in v1. Use ABI v2 instead where possible.

Structs

ChipInfo

Information about a particular GPIO chip.

EventRequest

Information about a GPIO event request.

EventRequestFlags

Additional configuration flags for event requests.

HandleConfig

Updated configuration for an existing GPIO handle request.

HandleRequest

Information about a GPIO line handle request.

HandleRequestFlags

Configuration flags for requested lines.

LineEdgeEvent

Information about an edge event on a requested line.

LineInfo

Information about a certain GPIO line.

LineInfoChangeEvent

Information about a change in status of a GPIO line.

LineInfoFlags

Flags indicating the configuration of a line.

LineValues

The logical values of the requested lines.

Offsets

A collection of line offsets.

ValidationError

A failure to validate a struct returned from a system call.

Enums

LineEdgeEventKind

The trigger identifier for a LineEdgeEvent.

LineInfoChangeKind

The trigger identifier for a LineInfoChangeEvent.

Functions

get_chip_info

Get the publicly available information for a chip.

get_line_event

Request a line with edge detection enabled.

get_line_handle

Request a line or set of lines for exclusive access.

get_line_info

Get the publicly available information for a line.

get_line_values

Read the values of requested lines.

set_line_config

Update the configuration of an existing handle or event request.

set_line_values

Set the values of requested lines.

unwatch_line_info

Remove any watch on changes to the LineInfo for a line.

watch_line_info

Add a watch on changes to the LineInfo for a line.

Type Definitions

Offset

An identifier for a line on a particular chip.