Module gpiocdev_uapi::v2

This module implements GPIO ABI v2 which is the current version of the ABI, released in Linux v5.10.

Structs

ChipInfo

Information about a particular GPIO chip.

LineAttribute

A configurable attribute of a line.

LineAttributes

The set of potential configuration attributes for a line.

LineConfig

Configuration for a set of requested lines.

LineConfigAttribute

A configuration attribute associated with one or more of the requested lines.

LineConfigAttributes

The set of additional configuration attributes for a line request.

LineEdgeEvent

Information about an edge event on a requested line.

LineFlags

Flags indicating the configuration of a line.

LineInfo

Information about a certain GPIO line.

LineInfoChangeEvent

An event indicating a change to the info for a line.

LineRequest

Information about a request for GPIO lines.

LineValues

Values of GPIO lines.

Offsets

A collection of line offsets.

ValidationError

A failure to validate a struct returned from a system call.

Enums

LineAttributeKind

An identifier for which field of the LineAttributeValueUnion is in use.

LineAttributeValue

The attribute value contained within a LineAttribute.

LineEdgeEventKind

The trigger identifier for a LineEdgeEvent.

LineInfoChangeKind

The trigger identifier for a LineInfoChangeEvent.

Constants

NUM_ATTRS_MAX

The capacity of LineAttributes and LineConfigAttributes arrays.

Functions

get_chip_info

Get the publicly available information for a chip.

get_line

Request a line or set of lines for exclusive access.

get_line_info

Get the publicly available information for a line.

get_line_values

Read values of requested lines.

set_line_config

Update the configuration of an existing line request.

set_line_values

Set values of requested output lines.

unwatch_line_info

Remove any watch on changes to the LineInfo for a line.

watch_line_info

Add a watch on changes to the LineInfo for a line.

Type Definitions

Offset

An identifier for a line on a particular chip.

Unions

LineAttributeValueUnion

The value of a particular line attribute.